apps/backend/requirements.txt
fastapi==0.115.12
uvicorn[standard]==0.34.0
python-multipart==0.0.20

# Document Parsers
pandas==2.2.3
openpyxl==3.1.5
python-pptx==1.0.2
PyMuPDF==1.25.5

# AI / Embeddings
sentence-transformers==3.4.1
chromadb==0.6.3
langchain-text-splitters==0.3.8
openai>=1.55.0

# Utils
python-dotenv==1.1.0
aiofiles==24.1.0
pydantic==2.11.3
